requirements.txt
#Server
Django==1.4.1
gunicorn==0.14.6
psycopg2==2.4.5
dj-database-url==0.2.0
wsgiref==0.1.2
gevent==0.13.4
South==0.7.4

celery>=2.3.3
django-celery==3.0.10
